datashufflepy-zeus/requirements.txt
#源 -i http://pypi.douban.com/simple/ --trusted-host pypi.douban.com
arrow==0.14.2
pymongo==3.8.0
JPype1-py3==0.5.5.4
JayDeBeApi==1.1.1
PyMySQL==0.9.3
xlrd==1.2.0
retrying==1.3.3
requests==2.22.0
pdfminer3k==1.3.1
PyPDF2==1.26.0
